Broke cameras into individual modules
